pyproject.toml
[tool.poetry]
name = "www-bareasgi-com"
version = "0.1.0"
description = "The bareASGI web site"
authors = ["Rob Blackbourn <rob.blackbourn@googlemail.com>"]
license = "Apache-2.0"
packages = [
    { include = "bareasgi_web_site" }
]

[tool.poetry.scripts]
start = "bareasgi_web_site.server:start_server"

[tool.poetry.dependencies]
python = "^3.7"
hypercorn = "^0.9.0"
uvloop = "^0.14.0"
PyYaml = "^5.1"
markdown = "^3.1"
bareasgi = "^3.4"
pygments = "^2.4"
bareasgi-static = "^3.0"

[tool.poetry.dev-dependencies]
mypy = "^0.740.0"
autopep8 = "^1.4"
pylint = "^2.4"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
